testing the submission process
