+++ index.html
@@ -6,7 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Pokemon Lab</title>
     <!-- <link rel="stylesheet" href="css/styles.css"> -->
-    <link rel="stylesheet" href="css/pokemon-table.css"> --> -->
+    <!-- <link rel="stylesheet" href="css/pokemon-table.css"> --> --> -->
 </head>
 <body>
     <header>
@@ -46,8 +46,8 @@
 
     </main>
 
-    <script type="module" src="js/data/app.js"></script>
-    <script type="module" src="js/data/pokedex-api.js"></script>
-    <script type="module" src="js/data/pokedex-table.js"></script>
+    <script type="module" src="js/app.js"></script>
+    <script type="module" src="js/pokedex-api.js"></script>
+    <script type="module" src="js/pokedex-table.js"></script>
 </body>
 </html>
