--- pokemon-table.css
@@ -1,3 +1,0 @@
-img {
-width: 50px
-}
