--- styles.css
@@ -1,54 +1,0 @@
-body {
-    font-family: Corbel, "Lucida Grande", "Lucida Sans Unicode", "Lucida Sans", "DejaVu Sans", "Bitstream Vera Sans", "Liberation Sans", Verdana, "Verdana Ref", sans-serif;
-    color: #000000;
-    margin: 0;
-}
-
-* {
-    box-sizing: border-box;
-}
-
-
-
-.action {
-    background: #39499B;
-    color: white;
-    border: 1px solid white;
-}
-
-.secondary {
-    background: #00AA8F;
-    color: white;
-}
-
-.caution {
-    background: #EFEB3E;
-    color: black;
-    border: 1px solid black;
-}
-
-.danger {
-    color: #EB4D5C;
-}
-
-header {
-    background: black;
-    color: white;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-
-img, header {  
-    height: 100%;
-    margin-right: 15px;
-} 
-
-header, section {
-    padding: 10px 20px;
-}
-
-h1 {
-    margin: 0;
-}
